backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pypdf
spacy
scikit-learn
sentence-transformers
httpx
numpy
pandas
matplotlib
seaborn
jinja2
python-slugify
